usb: xhci: replace custom ring segment list implementation with list.h
Replace the custom list implementation with the standard list.h.
The transition is made with minimal function adjustments to maintain the
simplicity and readability of this commit. Functions will be adjusted and
optimized for the new list style in future commits.

The reason for this change is to simplify xhci ring segment by using a
common list system, as well as enabling us to use the features provided by
list.h, such as bilateral list access.

Add help functions _get_next_deq_seg() and _get_next_enq_seg(). These
functions help keep the code clutter free and simple.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>