usb: xhci: replace custom ring segment list with standard list.h lists
Replace custom list with list.h list with minimal amount of function
adjustments. Functions will be adjusted/optimized for the new list
style in future commits. This is to not over complicate this commit.

The reason for this change is to simplify xhci ring segment by using
a common list system, as well as enabling us to use the features
provided by list.h, such as bilateral list access.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>