xhci: remove struct xhci_ring's pointers to first & last segment
Instead use 'RING_FIRST_SEG()' and 'RING_LAST_SEG()' macros, which use
the list.h macro for getting the first and last entry from list.
This approach reduces the risk that the first/last segment pointers
are incorrect.

Reason for introducing the xHCI specific macros instead of using the
list.h macro directly, is that the list.h macro requires the data type
and name of the 'list_head' variable. Thus, it is more user friendly
and cleaner to hide them in the xHCI macro.

Example:
	RING_FIRST_SEG(&ring->seg_list);
VS
	list_first_entry(&ring->seg_list, struct xhci_segment, list);

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>