usb: xhci: centralize default interrupt state initialization
Move all interrupt initialization into a single function and rename it to
"xhci_init_interrupter". This centralization helps keep all relevant code
in one place.

Function xhci_run() is called for both the primary and secondary HCD,
while xhci_init_interrupter() is only called for the primary HCD.
This means that IP auto-clear will only be set for the primary HCD,
which is optimal, instead of being set twice.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>