usb: xhci: integrate 'COMP_SUCCESS' check into switch case
Instead of having an if-else structure with a switch case inside the else
statement, integrate the 'if (cmd_comp_code == COMP_SUCCESS)' condition
into the switch case. This way, all command completion codes are handled
within a single switch case.

This refactor simplifies and cleans up the code, without any functional
changes.

This patch is in preparation for the xhci_handle_cmd_set_deq() rework.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>