usb: xhci: improve xhci_clear_interrupt_pending()
The function appears straightforward but has several nuances.
* Add comments to explain its quirks and provide better clarity.
* Add read, to flush the buffer.

xHCI specification 1.2, section 5.5.2.1 "Note" states:
"In systems that do not support MSI or MSI-X, the IP bit may be cleared
by writing a '1' to it. Most systems have write buffers that minimize
overhead, but this may require a read operation to guarantee that the
write has been flushed from posted buffers."

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>