usb: xhci: handle Set TR Deq Error due to Disabled slot state
One of the reasons for a Set TR Deq Context State Error, is that the Slot
state is not Default, Configured, or Addressed. This leaves it 2 options;
Enabled and Disabled. If the slot state was Enabled, HW would have returned
a Slot Not Enabled Error, which only leaves Disabled Slot for Set TR Deq
Context State Error.

Due to the complexity of transitioning the Device Slot from Disabled ->
Enabled -> Default, this specific Context State Error case isn't corrected.

xHCI specification, rev 1.2:
 Slot State		section 4.5.3
 Set TR Dequeue Pointer	section 4.6.10

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>