usb: xhci: simplify Set TR Deq command desync
The hardware might return a successful Set TR Deq command completion
event, yet the hardware dequeue pointer may not match the xhci driver
dequeue pointer. This desynchronization indicates that the hardware and
the xhci driver are out of sync, which is a severe error.

Currently, there is no handling for Set TR Deq desync. Instead, a warning
is printed, and the xhci driver proceeds as if the pointers are in sync,
causing the xhci driver to fail eventually.

Due to the severity and complexity of handling this issue, the function
will now return immediately upon detecting desynchronization, preventing
further complications. This change simplifies future work on
xhci_handle_cmd_set_deq() by not attempting to accommodate this anomaly.

Additionally, change the warnings to error messages, as this issue will
lead to failure.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>